Update README.md
The minimally invasive surgery has emerged as a safe alternative to traditional open surgeries, and have become the first preference of many doctors today. The minimally invasive surgery, introduced to the medical world in the 1980s, holds relatively fewer risks, and often fast tracks the recovery rate. These techniques are thus safer than the traditional methods and are all the more compelling.